requirements.txt
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
psycopg2==2.8.5
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.1.0